Job CRUD ok (avec l'utilisation d'un default user). Finalisation: nettoyage du code, vérification de l'ordre des tabulation...

+++ Connect/Connect/ListingEtudiant.cs
@@ -35,14 +35,18 @@ namespace Connect
             
             Int32.TryParse(comboBoxListEtudiant.SelectedValue.ToString(), out value);
             etudiantSelectionne = EtudiantManager.GetEtudiant(value);
-            if (value != -1)
+            if (value != -1 && value != 1)
             {
                 FormEtudiant formEtudiant = new FormEtudiant(etudiantSelectionne);
                 formEtudiant.MdiParent = this.MdiParent;
                 formEtudiant.Show();
                 this.Close();
             }
+            // Protection de l'id de l'étudiant par défaut
+            if(value == 1)
+                MessageBox.Show("Ligne protégée pour la stabilité de l'application", "Attention", MessageBoxButtons.OK, MessageBoxIcon.Warning);
         }
+
         /// <summary>
         /// Permet la suppression d'un profil étudiant et de ses périodes de disponibilité
         /// </summary>
@@ -52,60 +56,72 @@ namespace Connect
         {
             int value;
             Int32.TryParse(comboBoxListEtudiant.SelectedValue.ToString(), out value);
-            
-            var Result = MessageBox.Show("Etes-vous sûr de vouloir supprimer l'étudiant n°" + value.ToString() + "?", "Veuillez confirmer:", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning);
-            if (Result == DialogResult.OK)
+
+            // Protection de l'id de l'étudiant par défaut
+            if (value == 1)
+                MessageBox.Show("Ligne protégée pour la stabilité de l'application", "Attention", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+            else
             {
-                // la variable checkJob va permettre de conserver l'information booléenne quant aux attributions de job de l'étudiant en dehors de la boucle foreach
-                bool checkJob = false;
-                Connectds.jobDataTable jobDT = EntrepriseManager.GetJobDT();
-                foreach (Connectds.jobRow jobRow in jobDT)
+                var Result = MessageBox.Show("Etes-vous sûr de vouloir supprimer l'étudiant n°" + value.ToString() + "?", "Veuillez confirmer:", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning);
+                if (Result == DialogResult.OK)
                 {
-                    if (value == jobRow.etudiant_id)
+                    // la variable checkJob va permettre de conserver l'information booléenne quant aux attributions de job de l'étudiant en dehors de la boucle foreach
+                    bool checkJob = false;
+                    Connectds.jobDataTable jobDT = EntrepriseManager.GetJobDT();
+                    foreach (Connectds.jobRow jobRow in jobDT)
                     {
-                        var Result2 = MessageBox.Show("L'étudiant n°" + value.ToString() + " a des attributions de jobs actives. Vous ne pouvez le supprimer. Souhaitez-vous rendre son profil inactif?", "Veuillez préciser:", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
-                        if (Result2 == DialogResult.Yes)
+                        if (value == jobRow.etudiant_id)
                         {
-                            etudiantSelectionne = EtudiantManager.GetEtudiant(value);
-                            etudiantSelectionne.statut_etudiant = false; // le statut devient inactif
-                            EtudiantManager.SaveEtudiant(etudiantSelectionne);
-                            PopulateAndBind();
-                        }
-                        checkJob = true;
-                        break;
-                    }
-                }
-                if (!checkJob)
-                {
-                    List<Connectds.periodeRow> periodeList = new List<Connectds.periodeRow>();
-                    periodeList = EtudiantManager.GetPeriodeList(value);
-                    if (periodeList != null)
-                    {
-                        foreach (var periode in periodeList)
-                        {
-                            EtudiantManager.DeletePeriode(periode.periode_id);
+                            var Result2 = MessageBox.Show("L'étudiant n°" + value.ToString() + " a des attributions de jobs actives. Vous ne pouvez le supprimer. Souhaitez-vous rendre son profil inactif?", "Veuillez préciser:", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+                            if (Result2 == DialogResult.Yes)
+                            {
+                                etudiantSelectionne = EtudiantManager.GetEtudiant(value);
+                                etudiantSelectionne.statut_etudiant = false; // le statut devient inactif
+                                EtudiantManager.SaveEtudiant(etudiantSelectionne);
+                                PopulateAndBind();
+                            }
+                            checkJob = true;
+                            break;
                         }
                     }
-                    EtudiantManager.DeleteEtudiant(value);
-                    PopulateAndBind();
+                    if (!checkJob)
+                    {
+                        List<Connectds.periodeRow> periodeList = new List<Connectds.periodeRow>();
+                        periodeList = EtudiantManager.GetPeriodeList(value);
+                        if (periodeList != null)
+                        {
+                            foreach (var periode in periodeList)
+                            {
+                                EtudiantManager.DeletePeriode(periode.periode_id);
+                            }
+                        }
+                        EtudiantManager.DeleteEtudiant(value);
+                        PopulateAndBind();
+                    }
                 }
-            }
+            } 
         }
 
         private void buttonDisponibilite_Click(object sender, EventArgs e)
         {
             int value;
             Int32.TryParse(comboBoxListEtudiant.SelectedValue.ToString(), out value);
-            this.Close();
-            Disponibilite disponibilite = new Disponibilite(value);
-            disponibilite.MdiParent = HomePage.ActiveForm;
-            disponibilite.Show();
+
+            // Protection de l'id de l'étudiant par défaut
+            if (value == 1)
+                MessageBox.Show("Ligne protégée pour la stabilité de l'application", "Attention", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+            else
+            {
+                this.Close();
+                Disponibilite disponibilite = new Disponibilite(value);
+                disponibilite.MdiParent = HomePage.ActiveForm;
+                disponibilite.Show();
+            }
         }
 
         private void buttonListEtudiantFermer_Click(object sender, EventArgs e)
         {
             this.Close();
         }
-
     }
 }
